Adding the Complaint Book page, both front and back

+++ core/database/migrations/2024_04_07_105257_create_sessions_table.php
@@ -4,16 +4,20 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 class CreateSessionsTable extends Migration{
   public function up(){
-    Schema::create('sessions', function (Blueprint $table) {
-      $table->string('id')->unique();
-      $table->foreignId('user_id')->nullable();
-      $table->string('ip_address', 45)->nullable();
-      $table->text('user_agent')->nullable();
-      $table->text('payload');
-      $table->integer('last_activity');
-    });
+    if(!Schema::hasTable('sessions')){
+      Schema::create('sessions', function (Blueprint $table) {
+        $table->string('id')->unique();
+        $table->foreignId('user_id')->nullable();
+        $table->string('ip_address', 45)->nullable();
+        $table->text('user_agent')->nullable();
+        $table->text('payload');
+        $table->integer('last_activity');
+      });
+    }
   }
   public function down(){
-    Schema::dropIfExists('sessions');
+    if(Schema::hasTable('sessions')){
+      Schema::dropIfExists('sessions');
+    }
   }
 }
